fix: upd.  osCommands, hash, compress, decompress commands.
