+++ src/app.js
@@ -1,71 +1,11 @@
 import { createInterface } from "readline/promises";
 import { cwd, chdir, exit } from "process";
-import {
-  readdirSync,
-  createReadStream,
-  createWriteStream,
-  mkdirSync,
-  unlinkSync,
-  readFileSync,
-  existsSync,
-} from "fs";
-import { join, dirname, parse, basename } from "path";
-import os from "os";
-import { createHash } from "crypto";
-import { createGzip, createUnzip } from "zlib";
 import { tryCatchWrapper } from "./utils/utils.js";
-import { add, cat, cp, ls, mv, rm, rn } from "./utils/helpers.js";
+import { add, cat, compress, cp, decompress, hash, ls, mv, osCommands, rm, rn } from "./utils/helpers.js";
 
 export class App {
   constructor(startDir) {
     this._curentPath = startDir;
-  }
-
-  async compress(src, dest) {
-    const gzip = createGzip();
-    readdirSync(cwd()).map((el) => {
-      if (el === src) {
-        const rs = createReadStream(src, { encoding: "utf-8" }).on(
-          "error",
-          (err) => {
-            console.log("error rs", err);
-          }
-        );
-        if (!existsSync(dest)) {
-          !!Object.values(parse(dest))[1] &&
-            mkdirSync(dirname(dest), { recursive: true });
-          const ws = createWriteStream(`${dest}`)
-            .on("error", (err) => console.log("err ws", err))
-            .on("finish", () => console.log("compressed successfully"));
-          rs.pipe(gzip).pipe(ws);
-        }
-      }
-    });
-  }
-  async decompress (src, dest) {
-    const unZip = createUnzip();
-    readdirSync(cwd()).map((el) => {
-      if (el === src) {
-        if (parse(src).ext === ".gz") {
-          const rs = createReadStream(basename(src)).on(
-            "error",
-            (err) => {
-              console.log(err);
-            }
-          );
-          if (!existsSync(dest)) {
-            !!Object.values(parse(dest))[1] &&
-              mkdirSync(dirname(argTwo), { recursive: true });
-            const ws = createWriteStream(`${dest}`)
-              .on("error", (err) => console.log("err ws", err))
-              .on("finish", () => {
-                console.log("unziped succeffully");
-              });
-            rs.pipe(unZip).pipe(ws);
-          }
-        }
-      }
-    });
   }
 
   async start() {
@@ -85,7 +25,6 @@ export class App {
         case ".quit":
         case ".q":
           exit();
-          break;
         case "up":
           tryCatchWrapper(chdir, '..');
           break;
@@ -114,45 +53,16 @@ export class App {
           tryCatchWrapper(rm, arg);
           break;
         case "os":
-          switch (arg) {
-            case "--EOL":
-              const input = JSON.stringify(os.EOL);
-              console.log(
-                "end-of-line marker is :",
-                input.substring(1, input.length - 1)
-              );
-              break;
-            case "--cpus":
-              const information = [
-                {
-                  cpus: os.cpus().length,
-                  model: os.cpus()[0].model,
-                },
-              ];
-              console.table(information);
-              break;
-            case "--homedir":
-              console.log(os.homedir());
-              break;
-            case "--username":
-              console.log(os.userInfo().username);
-              break;
-            case "--architecture":
-              console.log(os.arch());
-              break;
-            default:
-              console.log('Invalid input');
-              break;
-            }
+         tryCatchWrapper(osCommands, arg);
             break;
             case "hash":
-              await this.hash(arg);
+         tryCatchWrapper(hash, arg);
           break;
           case "compress":
-            await this.compress(arg, argTwo);
+            compress(arg, argTwo);
             break;
             case "decompress":
-              await this.decompress(arg, argTwo);
+              decompress(arg, argTwo);
               break;
               default:
                 console.log('Invalid input');
